requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
schedule==1.2.0
python-dotenv==1.0.0
pytz==2023.3
newspaper3k==0.2.8
lxml==4.9.3
lxml-html-clean==0.1.0
nltk==3.8.1
tldextract==5.1.1 